serveral changes to file structure and server
