--- server/graphql/resolvers/User/index.js
@@ -1,7 +1,0 @@
-const query = require("./query");
-const mutation = require("./mutation");
-
-module.exports = {
-  UserQuery: query,
-  UserMutation: mutation,
-};
