--- server/graphql/resolvers/User/query.js
@@ -1,15 +1,0 @@
-const models = require("../../../models");
-
-async function getUser(root, args, { user }) {
-  try {
-    if (!user) throw new Error("You are not authenticated!");
-    // TODO: update later
-    return await models.User.findAll({ username: "abc" });
-  } catch (error) {
-    throw new Error(error.message);
-  }
-}
-
-module.exports = {
-  getUser,
-};
