--- server/graphql/resolvers/index.js
@@ -1,14 +1,0 @@
-const { UserQuery, UserMutation } = require("./User");
-const { PlaceMutation } = require("./Place");
-
-const resolvers = {
-  Query: {
-    ...UserQuery,
-  },
-  Mutation: {
-    ...UserMutation,
-    ...PlaceMutation,
-  },
-};
-
-module.exports = resolvers;
