--- server/graphql/typedefs/index.graphql
@@ -1,24 +1,0 @@
-type User {
-  user_id: String!
-  username: String!
-  fullname: String!
-  password: String!
-}
-
-type AuthPayload {
-  token: String!
-  user: User!
-}
-
-type Query {
-  getUser: [User]
-}
-
-type Mutation {
-  register(
-    username: String!
-    fullname: String!
-    password: String!
-  ): AuthPayload!
-  login(username: String!, password: String!): AuthPayload!
-}
