+++ server/schemas/typeDefs.js
@@ -1,7 +1,33 @@
 const { gql } = require('apollo-server-express');
 
 const typeDefs = gql`
-    type Comment {
+
+type User {
+    user_id: String!
+    username: String!
+    fullname: String!
+    password: String!
+  }
+  
+  type AuthPayload {
+    token: String!
+    user: User!
+  }
+  
+  type Query {
+    getUser: [User]
+  }
+  
+  type Mutation {
+    register(
+      username: String!
+      fullname: String!
+      password: String!
+    ): AuthPayload!
+    login(username: String!, password: String!): AuthPayload!
+  }
+
+  type Comment {
     _id: ID!
     commentBody: String!
     username: String!
@@ -35,5 +61,4 @@ const typeDefs = gql`
         removeComment(commentId: ID!): User
     }
 `;
-
 module.exports = typeDefs;
