+++ server/server.js
@@ -1,40 +1,46 @@
-const { ApolloServer } = require("apollo-server-express");
+const express = require('express');
+const { ApolloServer } = require('apollo-server-express');
+const path = require('path');
+const { authMiddleware } = require('./utils/auth');
 
-const fs = require("fs");
-const path = require("path");
-require("dotenv").config();
-require("./config/database").connect();
+const { typeDefs, resolvers } = require('./schemas');
+const db = require('./config/connection');
 
-const app = require("express")();
-
-const port = process.env.PORT || 3000;
-
-const gqlFiles = fs.readdirSync(path.join(__dirname, "./graphql/typedefs"));
-let typeDefs = "";
-gqlFiles.forEach((file) => {
-  typeDefs += fs.readFileSync(
-    path.join(__dirname, "./graphql/typedefs", file),
-    {
-      encoding: "utf8",
-    }
-  );
+const PORT = process.env.PORT || 3001;
+const app = express();
+const server = new ApolloServer({
+  typeDefs,
+  resolvers,
+  context: authMiddleware,
 });
 
-const resolvers = require("./graphql/resolvers");
+app.use(express.urlencoded({ extended: false }));
+app.use(express.json());
 
-let apolloServer = {
-  graphqlPath: "",
-};
+// Serve up static assets
+// app.use('/images', express.static(path.join(__dirname, '../client/images')));
 
-async function startServer() {
-  apolloServer = new ApolloServer({ typeDefs, resolvers });
-  await apolloServer.start();
-  apolloServer.applyMiddleware({ app, path: "/graphql" });
+if (process.env.NODE_ENV === 'production') {
+  app.use(express.static(path.join(__dirname, '../client/build')));
 }
-startServer();
 
-app.listen(port, () => {
-  console.log(
-    `🚀 Server ready at http://localhost:${port}${apolloServer.graphqlPath}`
-  );
+app.get('/', (req, res) => {
+  res.sendFile(path.join(__dirname, '../client/build/index.html'));
 });
+
+
+// Create a new instance of an Apollo server with the GraphQL schema
+const startApolloServer = async () => {
+  await server.start();
+  server.applyMiddleware({ app });
+  
+  db.once('open', () => {
+    app.listen(PORT, () => {
+      console.log(`API server running on port ${PORT}!`);
+      console.log(`Use GraphQL at http://localhost:${PORT}${server.graphqlPath}`);
+    })
+  })
+  };
+  
+// Call the async function to start the server
+  startApolloServer();
